* Make CMakeLists support more than just g++ * Remove use of sys/time as it is POSIX compliance dependant and this is a multiplatform server * Substitute in std::chrono as we are using C++11 anyway * Include algorithm in Object.cpp as max isn't present without it (and remove the "std::" prefix as that file has a using namespace std anyway) * Undef the min macro in PacketHandler.cpp as windows specific libraries define it

+++ gamed/include/Map.h
@@ -20,7 +20,7 @@ private:
 public:
    Map(Game* game) : game(game) { }
    virtual ~Map() { }
-   virtual void update(unsigned int diff);
+   virtual void update(long long diff);
    Object* getObjectById(uint32 id);
    void addObject(Object* o);
